Improve type_check_failed exception
type_check_exception is thrown when an optional parameter is passed in a
GCM notification, and the parameter fails a type check.

For example, collapse_key must be a binary, and is passed a boolean
(say), gcm_json:make_notification/1 throws the exception.

The problem is that there is not enough information in the exception
to diagnose what the issue is without poring over source code.

As an example, consider this exception:

{type_check_failed,{#Fun<gcm_json.3.38571724>,true}}

All we can tell from this is that a type check failed due to some fun in
gcm_json, which is not very helpful.

This commit changes the exception to include the name of the predicate,
the key, and the value.

The same exception as above now produces this:

{type_check_failed,{pred,is_binary,key,collapse_key,val,true}}}

This makes it easy to see that 'collapse_key' was passed in the
notification as 'true', when it should have been a binary.