Add asynchronous interface, and refactor
- Support 'to' in notification as well as 'id'.
- Fix stop/1 and other bugs.
- Add and modify tests.

ISSUES
======
There is a real issue with this module as it stands.

GCM supports sending a notification to multiple registration ids,
and although this module supports sending the request, it doesn't
handle the multi-valued response.

Right now, it throws an exception on receiving a multi-valued
response, which is not great, but better than pretending everything
is ok.

We need to refactor APNS, GCM, sc_push, and the Web API, so that the
top level notification APIs are more consistent with each other.