WIP: Improve test coverage, fix bugs
There is a real issue with this module as it stands.

GCM supports sending a notification to multiple registration ids,
and although this module supports sending the request, it doesn't
handle the multi-valued response.

Right now, it throws an exception on receiving a multi-valued
response, which is not great, but better than pretending everything
is ok.

An effort needs to be made to refactor both the APNS and GCM
modules, as well as sc_push and the Web API, so that the results
of sending notifications at the top level look the same for both
push systems. Most of this code was written originally in 2012,
when GCM and APNS worked very differently. Now that GCM also
supports APNS, it might be time to change the interfaces to look
more like GCM and achieve consistency in that way.